+++ node/binding/src/lib.rs
@@ -16,7 +16,7 @@ use swc_common::{
     FilePathMapping, SourceMap,
 };
 
-mod babel_ast;
+mod babelify;
 mod bundle;
 mod parse;
 mod print;
